Expanded function development
Tried adding the function to drag and drop a excel spreadsheet into the "Upload" window and return the data in a new window as a tree format (This is temporary, as the program will show the manifest data in a grid format shaped as a linehaul (truck). Currently broken. Will find the broken part and continue to expand these functions for the next push.