New modules and changes to assigning logic
Overhaul of the assigning logic used to apply freight to the grid (load plan). Including adding size constraints to each grid square so long items that will span multiple grid squares will visually show. This is still a work in progress commit. Also changed the spreadsheet to represent depth as width because I think it is better understood as depth when looking at a load plan. Will finish correct the new bugs to the assigning logic so it is working and will then add and expand on further modules.