Fixed modules/functions to assign correctly
Fixed the functions in the modules, so that they assigned freight to the grid correctly, make sure that items that had lengths greater than 1.2m (standard pallet) where then spread across connecting spaces. Now working on adjusting the assigner function/module so that it assigns the freight from right to left as to simulate front and rear of the truck/trailer. As well as making sure that the tallest pallets and heaviest are loaded first.