Assigning logic fixed
Corrected the assigner modules logic. It now assigns freight from the right to the left. Still getting issues assigning to both sides of the grids, plus the first 3 squares on the driver and passenger side need fixing. Only tall stuff can be allocated to those as there is no mezz floor to support freight above.